src/python/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
pydantic==2.5.0
httpx==0.25.1
pytest==7.4.3
numpy==1.24.3
sentence-transformers==2.2.2
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
